requirements.txt
# by ComfyUI
torch==2.1.1
torchsde
torchvision==0.16.1
einops
transformers==4.26.0
safetensors>=0.3.0
aiohttp
pyyaml
Pillow
scipy
tqdm
psutil
kornia>=0.7.1
--extra-index-url https://download.pytorch.org/whl/cu120

# by diffusers
diffusers

# by secmi: https://github.com/jinhaoduan/SecMI-LDM
omegaconf
matplotlib
scikit-learn
